requirements.txt
langchain
langchain-groq
langchain-community
streamlit
python-dotenv
requests
mitreattack-python
google-search-results
